Fix/models and schema (#7)
* chore(makefile): add docker-rebuild command

This commit adds a `docker-rebuild` command to the Makefile.
This command rebuilds the Docker images and restarts all
services using `docker-compose up --build`. It also checks
for the existence of the `.env` file before running.

* refactor(api): update campaigns route path

The campaigns route path has been updated to include the "/api"
prefix. This change ensures consistency and aligns with the overall
API structure.

* refactor(campaign): use chi router for ID extraction

Replaces custom ID extraction with chi router's URLParam function
for cleaner and more robust ID retrieval. Removes the
extractIDFromPath function.

* refactor(campaign): simplify CampaignWithStats struct

The `UpdatedAt` field was removed from the `Campaign` struct, and
the `CampaignWithStats` struct was modified to include only the
necessary fields from the `Campaign` struct and the `Stats` field.
This change aims to improve code maintainability.

* refactor(customer): remove CreatedAt and UpdatedAt fields

The fields were removed as they are managed at the database level
and should not be part of the Customer struct.

* refactor(repository): remove updated_at from campaign model

The updated_at field was removed from the database queries and
campaign model as it wasn't being used. This simplifies the code and
removes unnecessary data.

* refactor(repository): remove created_at, updated_at from queries

The database schema was changed to remove the created_at and
updated_at fields.  This commit updates the queries to reflect
those changes.

* refactor(campaign): simplify campaign list response

Return a simplified list item instead of the full campaign model.
This reduces the amount of data transferred and simplifies the API.

* feat(campaigns): add CampaignListItem DTO

This commit introduces the CampaignListItem DTO to represent a simplified
campaign view in lists. This DTO includes only essential fields like
ID, Name, Channel, Status, and CreatedAt, improving performance when
fetching campaign lists. The CampaignListResult now uses this DTO.

* refactor(db): simplify customer and campaign tables

Removes `updated_at` columns and `created_at` index from the
`customers` and `campaigns` tables. Also removes the triggers to
update the timestamps on update events for these tables, as they're
not needed anymore. Keeps the trigger for outbound_messages table.